changing more things trying to get singleUser query to work.
